Expand information in README on original models
Clarify in the README.md what the purpose of the repository is (as a a companion to a published paper) and clarify that the model files were converted from published models and are not original model files. Also advises the reader to consult the original publications in order to obtain the original model files.